feat(new-trace): Simplifying file structure
